Only ``stream`` stdout/stderr on kubernetes

+++ lib/galaxy/jobs/command_factory.py
@@ -39,6 +39,7 @@ def build_command(
     create_tool_working_directory=True,
     remote_command_params=None,
     remote_job_directory=None,
+    stream_stdout_stderr=False,
 ):
     """
     Compose the sequence of commands necessary to execute a job. This will
@@ -94,7 +95,7 @@ def build_command(
 
     for_pulsar = 'script_directory' in remote_command_params
     if not for_pulsar:
-        commands_builder.capture_stdout_stderr('../outputs/tool_stdout', '../outputs/tool_stderr')
+        commands_builder.capture_stdout_stderr('../outputs/tool_stdout', '../outputs/tool_stderr', stream_stdout_stderr=stream_stdout_stderr)
 
     # Don't need to create a separate tool working directory for Pulsar
     # jobs - that is handled by Pulsar.
@@ -277,7 +278,10 @@ class CommandsBuilder:
     def append_commands(self, commands):
         self.append_command("; ".join(c for c in commands if c))
 
-    def capture_stdout_stderr(self, stdout_file, stderr_file):
+    def capture_stdout_stderr(self, stdout_file, stderr_file, stream_stdout_stderr=False):
+        if not stream_stdout_stderr:
+            self.append_command(f'> "{stdout_file}" 2> "{stderr_file}"', sep="")
+            return
         trap_command = """trap 'rm -f "$__out" "$__err"' EXIT"""
         if TRAP_KILL_CONTAINER in self.commands:
             # We need to replace the container kill trap with one that removes the named pipes and kills the container
